Change file structure, add project support
This commit adds the SpotiFeed and TicTacToe components created to 'hold' each project. In Layout.js, these components are utilized to open a modal containing more information about the project. The modals themselves are contained withing modals.js, which contains all of the specific information for each project.

A specific stylesheet was created to handle any project styling, making it a bit easier to reference for debugging.